+++ judge/src/main/kotlin/JacksonInit.kt
@@ -1,10 +1,8 @@
-import com.fasterxml.jackson.core.JsonGenerator
-import com.fasterxml.jackson.core.JsonProcessingException
-import com.fasterxml.jackson.databind.*
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.databind.module.SimpleModule
 import com.fasterxml.jackson.databind.util.StdDateFormat
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
-import java.io.IOException
 
 
 val jsonMapper = ObjectMapper().apply {
@@ -25,30 +23,3 @@ val jsonMapper = ObjectMapper().apply {
     registerModule(simpleModule)
 }
 
-internal class CoordKeyDeserializer : KeyDeserializer() {
-    private val regex = Regex("(\\d+)_(\\d+)")
-    private fun coordFrom(str: String): Coord? {
-        val rm = regex.matchEntire(str)
-        val x = rm?.destructured?.component1()
-        val y = rm?.destructured?.component2()
-        if (x != null && y != null)
-            return Coord(x = x.toInt(), y = y.toInt())
-
-        return null
-    }
-
-    @Throws(IOException::class, JsonProcessingException::class)
-    override fun deserializeKey(key: String, ctxt: DeserializationContext):
-            Any? = coordFrom(key)
-}
-
-internal class CoordKeySerializer : JsonSerializer<Coord>() {
-    override fun serialize(
-        value: Coord?,
-        gen: JsonGenerator?,
-        serializers: SerializerProvider?
-    ) {
-        if (value != null && gen != null)
-            gen.writeFieldName("${value.x}_${value.y}")
-    }
-}
